MDL-47026 gradereport_grader: Removing the tooltips
They are not serving the same purpose as the one they had before,
and they were overlapping with the content of the cell. Positioning
them somewhere else made them more confusing than helpful.

This issue is a part of the MDL-46658 Task.
This issue is a part of the MDL-25544 Epic.